Update passenger to 5.1.8
Case EA-6868: Updated passenger to the latest stable version (5.1.8).

Updated the PassengerDisableHtaccess patch to handle the following
directives that were previously not disallowed in the .htaccess files
even when the PassengerDisableHtaccess directive was enabled:

    * Directives that are handled via the 'DEFINE_DIR_STR_CONFIG_SETTER' macro:

        PassengerAppRoot
        UnionStationKey

    * Directives that are handled via the 'DEFINE_DIR_THREEWAY_CONFIG_SETTER'
      macro:

        PassengerResolveSymlinksInDocumentRoot
        PassengerAllowEncodedSlashes
        UnionStationSupport
        PassengerBufferResponse

    * Directives with custom handlers:

        PassengerBaseURI
        UnionStationFilter

Reimplemented the 'supress empty log messages' patch for the upstream changes.

Additionally, removed the PassengerAppGroupName patch as this has been applied
upstream.