Stop reading the application's REVISION file
Case SEC-312: Stop reading the REVISION file in the application
directory.

Previously, an application's REVISION file was read as the root
user - which allowed malicious users to access any file on the
system by creating a symlink to the REVISION file and using the
passenger control sockets (through the passenger-status binary, etc)
to read the content.