EA-9194: Update scl-ruby24-passenger from v6.0.4 to v6.0.6
Added needed cxxbuilder files and altered the
PassengerDisableHtaccess patch to not include the changes
to the auto-generated files.  The build process can generate
those with the inclusion of cxxbuilder.
Also updated the spec file to look for the source files
under the new directory that the SOURCE tar extracts
into.